# lite_agent_web/lib/lite_agent_openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "web",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/v1/workspace/add": {
      "post": {
        "summary": "创建工作空间",
        "deprecated": false,
        "description": "",
        "tags": [
          "WorkspaceController"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "工作空间名字",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "认证 Token",
            "required": true,
            "example": "Bearer {{token}}",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDataString",
                  "description": ""
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 0,
                      "message": "",
                      "data": ""
                    }
                  }
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/v1/workspace/list": {
      "get": {
        "summary": "当前账号加入的工作空间",
        "deprecated": false,
        "description": "",
        "tags": [
          "WorkspaceController"
        ],
        "parameters": [
          {
            "name": "adminFlag",
            "in": "query",
            "description": "只查管理的工作空间,传1",
            "required": false,
            "example": 0,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "认证 Token",
            "required": true,
            "example": "Bearer {{token}}",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDataListWorkSpaceVO",
                  "description": ""
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 0,
                      "message": "",
                      "data": [
                        {
                          "id": "",
                          "name": "",
                          "role": ""
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/v1/workspace/memberList": {
      "get": {
        "summary": "工作空间的成员列表",
        "deprecated": false,
        "description": "",
        "tags": [
          "WorkspaceController"
        ],
        "parameters": [
          {
            "name": "username",
            "in": "query",
            "description": "用户名",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "",
            "required": true,
            "example": 10,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "pageNo",
            "in": "query",
            "description": "",
            "required": true,
            "example": 0,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "Workspace-id",
            "in": "header",
            "description": "工作空间",
            "required": true,
            "example": "{{workspaceId}}",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "认证 Token",
            "required": true,
            "example": "Bearer {{token}}",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDataPageModelWorkspaceMemberVO",
                  "description": ""
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 0,
                      "message": "",
                      "data": {
                        "pageNo": 0,
                        "pageSize": 0,
                        "total": 0,
                        "list": [
                          {
                            "id": "",
                            "workspaceId": "",
                            "userId": "",
                            "name": "",
                            "email": "",
                            "role": 0,
                            "createTime": ""
                          }
                        ]
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/v1/workspace/member": {
      "post": {
        "summary": "新增工作空间成员",
        "deprecated": false,
        "description": "",
        "tags": [
          "WorkspaceController"
        ],
        "parameters": [
          {
            "name": "emails",
            "in": "query",
            "description": "邮箱",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "query",
            "description": "设定角色(0-普通成员,1-管理员,2-开发者)",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "认证 Token",
            "required": true,
            "example": "Bearer {{token}}",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Workspace-id",
            "in": "header",
            "description": "工作空间id",
            "required": true,
            "example": "{{workspaceId}}",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDataString",
                  "description": ""
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 0,
                      "message": "",
                      "data": ""
                    }
                  }
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/v1/workspace/member/{memberId}": {
      "delete": {
        "summary": "删除工作空间成员",
        "deprecated": false,
        "description": "",
        "tags": [
          "WorkspaceController"
        ],
        "parameters": [
          {
            "name": "memberId",
            "in": "path",
            "description": "成员id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "认证 Token",
            "required": true,
            "example": "Bearer {{token}}",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Workspace-id",
            "in": "header",
            "description": "工作空间id",
            "required": true,
            "example": "{{workspaceId}}",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDataString",
                  "description": ""
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 0,
                      "message": "",
                      "data": ""
                    }
                  }
                }
              }
            }
          }
        },
        "security": []
      },
      "put": {
        "summary": "修改成员角色",
        "deprecated": false,
        "description": "",
        "tags": [
          "WorkspaceController"
        ],
        "parameters": [
          {
            "name": "memberId",
            "in": "path",
            "description": "成员id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "query",
            "description": "角色",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "认证 Token",
            "required": true,
            "example": "Bearer {{token}}",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Workspace-id",
            "in": "header",
            "description": "工作空间id",
            "required": true,
            "example": "{{workspaceId}}",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDataString",
                  "description": ""
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 0,
                      "message": "",
                      "data": ""
                    }
                  }
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/v1/workspace/activateInfo": {
      "get": {
        "summary": "获取激活信息",
        "deprecated": false,
        "description": "",
        "tags": [
          "WorkspaceController"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "description": "激活链接中获取参数",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDataDict",
                  "description": ""
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 0,
                      "message": "",
                      "data": {
                        "": {}
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/v1/workspace/activateMember": {
      "post": {
        "summary": "通过激活链接加入(登录)",
        "deprecated": false,
        "description": "",
        "tags": [
          "WorkspaceController"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "description": "激活链接中获取参数",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "username",
            "in": "query",
            "description": "昵称",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "password",
            "in": "query",
            "description": "密码",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDataString",
                  "description": ""
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 0,
                      "message": "",
                      "data": ""
                    }
                  }
                }
              }
            }
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ResponseDataPageModelWorkspaceMemberVO": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "状态码"
          },
          "message": {
            "type": "string",
            "description": "响应信息"
          },
          "data": {
            "$ref": "#/components/schemas/PageModelWorkspaceMemberVO",
            "description": "响应数据对象"
          }
        }
      },
      "PageModelWorkspaceMemberVO": {
        "type": "object",
        "properties": {
          "pageNo": {
            "type": "integer",
            "description": "当前第几页"
          },
          "pageSize": {
            "type": "integer",
            "description": "一页多少条记录"
          },
          "total": {
            "type": "integer",
            "description": "全部数据量"
          },
          "list": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkspaceMemberVO",
              "description": "工作空间成员列表信息"
            },
            "description": "数据"
          }
        }
      },
      "WorkspaceMemberVO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "成员id"
          },
          "workspaceId": {
            "type": "string",
            "description": "空间id"
          },
          "userId": {
            "type": "string",
            "description": "成员用户id"
          },
          "name": {
            "type": "string",
            "description": "昵称"
          },
          "email": {
            "type": "string",
            "description": "成员账号"
          },
          "role": {
            "type": "integer",
            "description": "空间成员角色"
          },
          "createTime": {
            "type": "string",
            "description": "加入时间"
          }
        }
      },
...
      "ResponseDataDict": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "状态码"
          },
          "message": {
            "type": "string",
            "description": "响应信息"
          },
          "data": {
            "$ref": "#/components/schemas/Dict",
            "description": "cn.hutool.core.lang.Dict"
          }
        }
      },
      "Dict": {
        "type": "object",
        "properties": {
          "key": {
            "$ref": "#/components/schemas/key"
          }
        }
      },
      "key": {
        "type": "object",
        "properties": {}
      },
      "ResponseDataListWorkSpaceVO": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "状态码"
          },
          "message": {
            "type": "string",
            "description": "响应信息"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkSpaceVO",
              "description": ""
            },
            "description": "响应数据对象"
          }
        }
      },
      "WorkSpaceVO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "工作空间id"
          },
          "name": {
            "type": "string",
            "description": "工作空间名字"
          },
          "role": {
            "type": "string",
            "description": "当前账号在里面的角色",
            "enum": [
              "0",
              "1",
              "2",
              "3",
              "ROLE_USER",
              "ROLE_ADMIN_DEL",
              "ROLE_DEVELOPER",
              "ROLE_ADMIN"
            ]
          }
        }
      },
      "ResponseDataString": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "状态码"
          },
          "message": {
            "type": "string",
            "description": "响应信息"
          },
          "data": {
            "type": "string",
            "description": "响应数据对象"
          }
        }
      }
    },
    "securitySchemes": {}
  },
  "servers": []
}